[refactor] request 요청 메소드 분리
